layout/generic/test/test_bug421839-2.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=421839
-->
<head>
  <title>Test for Bug 421839</title>
  <script type="application/javascript" src="/MochiKit/MochiKit.js"></script>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=421839">Mozilla Bug 421839</a>
<p id="display"><iframe src="bug421839-1-iframe.html"></iframe></p>
<div id="content" style="display: none">
  
</div>
<pre id="test">
<script type="application/javascript">

/** Test for Bug 421839 **/

function finish_421839() {
  ok(true, "pass");
  SimpleTest.finish();
}

setTimeout(finish_421839,2000)
SimpleTest.waitForExplicitFinish()


</script>
</pre>
</body>
</html>
